Knockback Script should be good to go
I managed to get the knockback script done, though keep in mind it's currently untested. It should work if I did it correctly however. The force of PlayerShoot and Range will definitely need to be played with. Also in case you do play with it, keep in mind that Shoot should be on the camera as that's where all the physics for the raycast is going to happen.